backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
firebase-admin==6.3.0
boto3==1.40.68
email-validator==2.3.0

